fix 4

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -12,16 +12,16 @@
 
 void print_binary(unsigned long int n)
 {
-	unsigned int temp = 1U << (sizeof(unsigned long int) * CHAR_BIT - 1);
+	unsigned int temp = 1UL << (sizeof(int) * CHAR_BIT - 1);
 	int leading_zeros = 0;
 	if (n == 0)
 	{
 		putchar('0');
 	}
 
-	while (temp != 0U)
+	while (temp != 0UL)
 	{
-		if ((temp & n) == 0U)
+		if ((temp & n) == 0UL)
 		{
 			if (leading_zeros == 1)
 				putchar('0');
